requirements.txt
anyio==4.11.0
asgiref==3.7.2
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
Django==4.2.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
firebase-admin==6.2.0
google-api-core==2.25.1
google-api-python-client==2.184.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
msgpack==1.1.1
pillow==10.2.0
proto-plus==1.26.1
protobuf==6.32.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
PyMuPDF==1.26.5
pyparsing==3.2.5
pytz==2023.3.post1
requests==2.32.5
rsa==4.9.1
setuptools==58.0.4
sniffio==1.3.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.2.0
urllib3==2.5.0
